Added option for to build to test site
If the user adds --test to the command line, then the region installs to /test this option will be helpful so that the manager can see the site, and notice if it contains errors before deploying to the actual site.  The --test flag is ignored when the --config option (for multiple sites) invoked.  Basically "test" is substituted in the code for the region name after the point in the code where the region repo is cloned from github.