Allow for builds without Visual Studio
A directory of configuration files required by MSBuild (targets) are
required for the build to succeed.  These targets are created when Visual
Studio 2012 is installed, but for scenarios where a different version (or
no version) is installed, the included directory contents can be copied
onto the system.  Additionally, nuget package restoring is automatically
turned on when the build is run, since the user cannot use VS to do so in
this circumstance.