Allow region_branch to override --dev
TNC uses --dev to auto install, but can't do so on region branches other
than 'development'.  Change the way that the options are applied so that
whenever specifying a region branch, that branch will be used. This is
implemented in an overly verbose way to avoid a confusing, nested
conditional.