Update git branch behavior
Makes the following changes to clarify branch usage

- Make the default framework branch develop, which will soon become the
default branch on Github
- Split the DEFAULT_BRANCH variable into DEFAULT_FRAMEWORK_BRANCH and
DEFAULT_REGION_BRANCH since they are slightly different (develop vs
development, respectively). Also note that on Github, the default branch
of the regions is often master. However, we want the default build
branch to be develop.
- Update --prod and --dev arguments to use master and develop. This was
the previous behavior, but it needed to be slightly updated to account
for the new default branches.

Also, update comments to remove references to installation, which does
not happen anymore.

Refs #36