Add support for building regions with story mode
If a region has store (a.k.a app intro) mode turned on, move around the
files and directories of the built region to support this mode. This
shuffling of files makes it so that the story mode app becomes the new
index for the region and the framework app can be embedded in the story
mode page. The filename conventions that make all this possible were
established by TNC developers, and must be correctly implemented for
story mode to work.

Refs #47